chore: move getDomainSeparator to common
